refactor: API 컨트롤러, 뷰 컨트롤러 분리
