+++ src/main/java/cart/controller/ViewController.java
@@ -3,22 +3,22 @@ package cart.controller;
 import cart.dto.response.ResponseMemberDto;
 import cart.dto.response.ResponseProductDto;
 import cart.service.CartService;
-import java.util.List;
-
 import cart.service.MemberService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 
+import java.util.List;
+
 @Controller
-public class CartController {
+public class ViewController {
 
     private final CartService cartService;
     private final MemberService memberService;
 
     @Autowired
-    public CartController(final CartService cartService, MemberService memberService) {
+    public ViewController(final CartService cartService, MemberService memberService) {
         this.cartService = cartService;
         this.memberService = memberService;
     }
@@ -35,5 +35,12 @@ public class CartController {
         final List<ResponseMemberDto> responseMemberDtos = memberService.findAll();
         model.addAttribute("members", responseMemberDtos);
         return "settings";
+    }
+
+    @GetMapping("/admin")
+    public String readAdminProducts(final Model model) {
+        final List<ResponseProductDto> responseProductDtos = cartService.findAll();
+        model.addAttribute("products", responseProductDtos);
+        return "admin";
     }
 }
