+++ src/main/java/cart/controller/exception/GlobalExceptionHandler.java
@@ -1,4 +1,4 @@
-package cart.controller;
+package cart.controller.exception;
 
 import java.util.HashMap;
 import java.util.List;
